requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
slowapi==0.1.9
openai==1.30.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
pdfplumber==0.11.1
numpy==1.26.4
